server/pyproject.toml
[project]
name = "happyword-server"
version = "0.5.1"
description = "Content backend for WordMagicGame (HarmonyOS NEXT client)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "motor>=3.6",
    # Beanie 1.27 calls list_collection_names with authorizedCollections=True
    # which mongomock-motor does not support; pin until upstream catches up.
    "beanie>=1.26,<1.27",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    # Use bcrypt directly; passlib is unmaintained and incompatible with bcrypt 5.x.
    "bcrypt>=4.2",
    "openai>=1.55",
    "httpx>=0.27",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "mongomock-motor>=0.0.34",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-python-jose>=3.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# Project rule: every commit must run pytest with 0 errors and 0 warnings.
# Default to errors so any new warning fails the suite. Narrow ignores below
# must always pin both the message prefix and the originating module, with a
# comment explaining why upstream cannot be fixed.
filterwarnings = [
    "error",
    # lazy_model 0.2.x (transitive dep of beanie<1.27) accesses
    # `instance.model_fields`, which Pydantic 2.11 deprecates. Beanie 1.27 fixes
    # this but is incompatible with mongomock-motor (drops the test runtime).
    # Re-evaluate when mongomock-motor catches up.
    "ignore:Accessing the 'model_fields' attribute on the instance is deprecated:DeprecationWarning:lazy_model",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "TCH"]

[tool.ruff.lint.flake8-bugbear]
# Depends() / Query() / Body() must be evaluated in argument defaults — that's the FastAPI idiom.
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.Query",
    "fastapi.Body",
    "fastapi.Path",
    "fastapi.Header",
    "fastapi.File",
    "fastapi.Form",
]

[tool.ruff.lint.per-file-ignores]
# Tests legitimately assert that *any* exception is raised when Beanie + mongomock
# disagree on the precise duplicate-key class.
"tests/*" = ["B017"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
